# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Текущий профиль",
        "tags": [
          "me"
        ]
      }
    },
    "/me/generations-log": {
      "get": {
        "operationId": "MeController_getGenerationsLog",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/avatar": {
      "post": {
        "operationId": "MeController_uploadAvatar",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/profile": {
      "patch": {
        "operationId": "MeController_updateProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/generations": {
      "post": {
        "operationId": "MeController_createGeneration",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-generations/{id}": {
      "delete": {
        "operationId": "MeController_deletePublicGeneration",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/generations/{id}/public": {
      "patch": {
        "operationId": "MeController_patchGenerationPublic",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/generations/{id}": {
      "delete": {
        "operationId": "MeController_deleteGeneration",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/tag": {
      "patch": {
        "operationId": "MeController_updateTag",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/wallet/topup": {
      "post": {
        "operationId": "MeController_walletTopup",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/wallet/promo": {
      "post": {
        "operationId": "MeController_walletPromo",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/wallet/roulette": {
      "post": {
        "operationId": "MeController_walletRoulette",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-portfolio": {
      "post": {
        "operationId": "MeController_addPortfolio",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-portfolio/{id}": {
      "delete": {
        "operationId": "MeController_deletePortfolio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/support/unread": {
      "get": {
        "operationId": "MeController_supportUnread",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Кол-во непрочитанных сообщений поддержки",
        "tags": [
          "me"
        ]
      }
    },
    "/me/support/messages": {
      "get": {
        "operationId": "MeController_supportMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "MeController_supportMessagesPost",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
    "/me/generations/reject": {
      "post": {
        "operationId": "MeController_rejectGeneration",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-feed": {
      "get": {
        "operationId": "MeController_myPublicFeed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-feed/{id}/like": {
      "post": {
        "operationId": "MeController_likeMyFeed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/public-feed/{id}/unlike": {
      "post": {
        "operationId": "MeController_unlikeMyFeed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/chats": {
      "get": {
        "operationId": "MeController_listChats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "MeController_createChat",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/chats/{chatId}": {
      "delete": {
        "operationId": "MeController_deleteChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "MeController_patchChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/chats/{chatId}/messages": {
      "get": {
        "operationId": "MeController_chatMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "MeController_postChatMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/me/chats/paid-mode/enable": {
      "post": {
        "operationId": "MeController_enablePaidMode",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "Neiro API",
    "description": "OpenAPI 3 спецификация, сгенерированная из NestJS-контроллеров (@nestjs/swagger).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Authorization: Bearer <token> (как в Express-сервере)"
      }
    },
...
  }
}